requirements.txt
flask
pillow
pytesseract
opencv-python-headless
torch
torchvision
